refactor: separated mapeditor into different files
